requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.3
PyMySQL==1.1.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
djangorestframework_simplejwt==5.3.1
pillow==10.2.0
